+++ src/terrain/PerlinNoise.cpp
@@ -12,16 +12,16 @@ void PerlinNoise::reseed(uint64_t seed) {
     noise->reseed(static_cast<uint32_t>(seed));
 }
 
-double PerlinNoise::evaluate(float x, float y) {
+double PerlinNoise::evaluate(double x, double y) {
     auto value = noise->accumulatedOctaveNoise2D(x / 1000.0, y / 1000.0, 8);
     return value;
 }
 
-double PerlinNoise::evaluate(float x, float y, float z) {
+double PerlinNoise::evaluate(double x, double y, double z) {
     return noise->noise3D(x, y, z);
 }
 
-glm::vec3 PerlinNoise::normal(float x, float y, double zoom) {
+glm::vec3 PerlinNoise::normal(double x, double y, double zoom) {
     x = x / 1000.0;
     y = y / 1000.0;
     auto dx = (evaluate((x + 1) / zoom, y / zoom) - evaluate(x / zoom, y / zoom)) / (1.0 / zoom);
